LPS-200485 Move tests to module
