+++ modules/apps/portal/portal-zip/src/test/java/com/liferay/portal/zip/internal/reader/ZipReaderImplTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.zip;
+package com.liferay.portal.zip.internal.reader;
 
 import com.liferay.petra.io.StreamUtil;
 import com.liferay.portal.kernel.test.util.DependenciesTestUtil;
