+++ modules/apps/portal/portal-zip/src/test/resources/com/liferay/portal/zip/internal/writer/dependencies/entry.txt
