+++ modules/apps/portal/source-formatter.properties
@@ -1,0 +1,6 @@
+##
+## Exclusions
+##
+
+    source.formatter.excludes=\
+        modules/apps/portal/portal-zip/src/test/resources/com/liferay/portal/zip/internal/reader/dependencies/file\.zip
